privatelink/requirements.txt
aws-cdk.aws-autoscaling
aws-cdk.aws-ec2
aws-cdk.aws-iam
aws-cdk.core
